QHash/QSet: Base iterator tag on warning version rather than deprecation
If you have deprecation warning version set to 5.15 and the
"disable deprecated before" set to something lower then it will complain
about usage of operator-- inside std::distance which is used in our
generic collection iterator.

This also required changing over to the QT_DEPRECATED_VERSION_5_15 macro
so that the deprecation warning would also be disabled.

This is a possible work-around - change the iterator tag if it would
trigger a warning.

Fixes: QTBUG-82397
Change-Id: I3e0ecae5edebba2a3560e7c3785bd9d1a6d0076d
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>